Rename globalconfig_types -> types
