+++ pkg/ddevapp/traefik_test.go
@@ -4,7 +4,7 @@ import (
 	"github.com/ddev/ddev/pkg/ddevapp"
 	"github.com/ddev/ddev/pkg/fileutil"
 	"github.com/ddev/ddev/pkg/globalconfig"
-	"github.com/ddev/ddev/pkg/globalconfig/globalconfigtypes"
+	"github.com/ddev/ddev/pkg/globalconfig/types"
 	"github.com/ddev/ddev/pkg/testcommon"
 	asrt "github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -27,7 +27,7 @@ func TestTraefikSimple(t *testing.T) {
 
 	ddevapp.PowerOff()
 	origRouter := globalconfig.DdevGlobalConfig.Router
-	globalconfig.DdevGlobalConfig.Router = globalconfigtypes.RouterTypeTraefik
+	globalconfig.DdevGlobalConfig.Router = types.RouterTypeTraefik
 	err = globalconfig.WriteGlobalConfig(globalconfig.DdevGlobalConfig)
 	require.NoError(t, err)
 	origConfig := *app
@@ -56,7 +56,7 @@ func TestTraefikSimple(t *testing.T) {
 	require.NoError(t, err, "failed to flush mutagen sync")
 
 	desc, err := app.Describe(false)
-	assert.Equal(desc["router"].(string), globalconfigtypes.RouterTypeTraefik)
+	assert.Equal(desc["router"].(string), types.RouterTypeTraefik)
 
 	// Test reachabiliity in each of the hostnames
 	httpURLs, httpsURLs, allURLs := app.GetAllURLs()
@@ -121,7 +121,7 @@ func TestTraefikVirtualHost(t *testing.T) {
 	require.NoError(t, err)
 
 	desc, err := app.Describe(false)
-	assert.Equal(globalconfigtypes.RouterTypeTraefik, desc["router"].(string))
+	assert.Equal(types.RouterTypeTraefik, desc["router"].(string))
 
 	// Test reachabiliity in each of the hostnames
 	httpURLs, _, allURLs := app.GetAllURLs()
