+++ pkg/globalconfig/types/types.go
@@ -1,4 +1,4 @@
-package globalconfigtypes
+package types
 
 type RouterType = string
