+++ pkg/versionconstants/versionconstants.go
@@ -2,9 +2,8 @@ package versionconstants
 
 import (
 	"fmt"
-	"github.com/ddev/ddev/pkg/globalconfig/globalconfigtypes"
-
 	"github.com/ddev/ddev/pkg/globalconfig"
+	"github.com/ddev/ddev/pkg/globalconfig/types"
 	"github.com/ddev/ddev/pkg/nodeps"
 )
 
@@ -100,7 +99,7 @@ func GetSSHAuthImage() string {
 // GetRouterImage returns the router image:tag reference
 func GetRouterImage() string {
 	image := TraefikRouterImage
-	if globalconfig.DdevGlobalConfig.Router == globalconfigtypes.RouterTypeNginxProxy {
+	if globalconfig.DdevGlobalConfig.Router == types.RouterTypeNginxProxy {
 		image = TraditionalRouterImage
 	}
 	return image
